Commit before Bug Test

+++ online_exam/exam/models/UserResultModel.py
@@ -5,6 +5,6 @@ class UserResult(models.Model):
     test_id = models.ForeignKey('exam.Test')
     test_type = models.IntegerField()
     gained_marks = models.IntegerField()
-    spend_time = models.IntegerField()
+    spend_time = models.DecimalField(max_digits=5, decimal_places=2)
     is_passed = models.BooleanField()
     datetime = models.DateTimeField()
